requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
psycopg[binary,pool]==3.2.6
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
email-validator==2.2.0
python-dotenv==1.0.1
fastapi-code-generator==0.5.4
